+++ Recalbox/userscripts/home_assistant_notifier.sh
@@ -4,7 +4,7 @@
 # A placer dans le dossier userscripts
 # Par Aurélien Tomassini
 
-SCRIPT_VERSION="home_assistant_notifier.sh:v1.5.4"
+SCRIPT_VERSION="home_assistant_notifier.sh:v1.5.5"
 
 # Configuration
 HOME_ASSISTANT_DOMAIN="homeassistant.local"